move timer and macro module
