+++ controller/subsystem_loader.go
@@ -9,15 +9,15 @@ import (
 
 	"github.com/reef-pi/reef-pi/controller/connectors"
 	"github.com/reef-pi/reef-pi/controller/equipment"
-	"github.com/reef-pi/reef-pi/controller/macro"
 	"github.com/reef-pi/reef-pi/controller/modules/ato"
 	"github.com/reef-pi/reef-pi/controller/modules/camera"
 	"github.com/reef-pi/reef-pi/controller/modules/doser"
 	"github.com/reef-pi/reef-pi/controller/modules/lighting"
+	"github.com/reef-pi/reef-pi/controller/modules/macro"
 	"github.com/reef-pi/reef-pi/controller/modules/ph"
 	"github.com/reef-pi/reef-pi/controller/modules/temperature"
+	"github.com/reef-pi/reef-pi/controller/modules/timer"
 	"github.com/reef-pi/reef-pi/controller/system"
-	"github.com/reef-pi/reef-pi/controller/timer"
 )
 
 func (r *ReefPi) loadPhSubsystem(bus i2c.Bus) error {
